+++ pa_dlna/tests/test_pa_dlna.py
@@ -477,7 +477,8 @@ class PatchSoapActionTests(IsolatedAsyncioTestCase):
         self.assertTrue(search_in_logs(logs.output, 'pa-dlna',
             re.compile("'remove' pulseaudio event .*previous state: running")))
         self.assertTrue(search_in_logs(logs.output, 'pa-dlna',
-            re.compile("'Stop' UPnP action .* device prev state: PLAYING")))
+            re.compile(
+                "'Closing-Stop' UPnP action .* device prev state: PLAYING")))
 
     async def test_prev_sink_input(self):
         ctx = PulseEventContext(sink_state='idle', sink_input_index=0)
@@ -528,26 +529,6 @@ class PatchSoapActionTests(IsolatedAsyncioTestCase):
             re.compile('change.* event .* prev/new state: running/running')))
         self.assertTrue(search_in_logs(logs.output, 'pa-dlna',
                 re.compile("MetaData\(.* artist='Ziggy Stardust'")))
-
-    async def test_ignored_SetAVTransportURI(self):
-        proplist = PROPLIST.copy()
-        proplist['media.title'] = 'Sticky Fingers'
-        ctx = PulseEventContext(prev_sink_state='running',
-                                sink_state='running',
-                                sink_input_index=1,
-                                sink_input_proplist=proplist)
-        self.assertTrue(ctx.renderer.nullsink.sink_input is None)
-
-        result, logs = await self.patch_soap_action('change', ctx,
-                                                    track_metadata=False)
-
-        self.assertTrue(ctx.renderer.nullsink.sink is ctx.sink)
-        self.assertTrue(ctx.renderer.nullsink.sink_input is ctx.sink_input)
-        self.assertEqual(len(result), 0)
-        self.assertTrue(search_in_logs(logs.output, 'pa-dlna',
-            re.compile('change.* event .* prev/new state: running/running')))
-        self.assertTrue(search_in_logs(logs.output, 'pa-dlna',
-            re.compile("'SetAVTransportURI' ignored UPnP action .* PLAYING")))
 
     async def test_soap_fault(self):
         ctx = PulseEventContext(sink_state='running', sink_input_index=0)